Refactor AnalyticsBaseModel and CommonModel classes
