+++ Common/AnalyticsModels/CommonModel.ts
@@ -14,7 +14,6 @@ import BadDataException from '../Types/Exception/BadDataException';
 import { JSONArray, JSONObject, JSONValue } from '../Types/JSON';
 import JSONFunctions from '../Types/JSONFunctions';
 import ObjectID from '../Types/ObjectID';
-import Text from '../Types/Text';
 
 export type RecordValue =
     | ObjectID
@@ -48,13 +47,7 @@ export default class CommonModel {
     }
 
     
-    private _tableName : string = '';
-    public get tableName() : string {
-        return this._tableName;
-    }
-    public set tableName(v : string) {
-        this._tableName = v;
-    }
+    
     
 
     public setColumnValue(
@@ -106,9 +99,8 @@ export default class CommonModel {
         throw new BadDataException('Column ' + columnName + ' does not exist');
     }
 
-    public constructor(data: { tableName: string,  tableColumns: Array<AnalyticsTableColumn> }) {
+    public constructor(data: { tableColumns: Array<AnalyticsTableColumn> }) {
         this.tableColumns = data.tableColumns;
-        this.tableName = data.tableName;
     }
 
     public getColumnValue<T extends RecordValue>(
@@ -243,7 +235,5 @@ export default class CommonModel {
         return json;
     }
 
-    public getAPIDocumentationPath(): string {
-        return Text.pascalCaseToDashes(this.tableName);
-    }
+    
 }
